move discovery components under P2P::Discovery
