+++ lib/ciri/p2p/server.rb
@@ -33,7 +33,7 @@ require_relative 'errors'
 require_relative 'peer'
 require_relative 'network_state'
 require_relative 'dialer'
-require_relative 'discovery_service'
+require_relative 'discovery/service'
 require_relative 'dial_scheduler'
 
 module Ciri
@@ -83,7 +83,7 @@ module Ciri
               task.sleep(0.5) until @local_address
 
               # start discovery service
-              @discovery_service = DiscoveryService.new(bootnodes: @bootnodes, host: @host, udp_port: @local_address.ip_port, tcp_port: @local_address)
+              @discovery_service = Discovery::Service.new(bootnodes: @bootnodes, host: @host, udp_port: @local_address.ip_port, tcp_port: @local_address)
               task.async { @discovery_service.run }
 
               # start dial outgoing nodes
